Criar categoria

+++ MyStore-backend/src/main/java/mystore/controllers/CategoriaController.java
@@ -1,8 +1,10 @@
 package mystore.controllers;
 
 import mystore.models.Categoria;
+import mystore.models.enums.RoleUtilizador;
 import mystore.services.CategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.persistence.EntityNotFoundException;
@@ -10,9 +12,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import static mystore.models.enums.RoleUtilizador.FUNCIONARIO;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.web.bind.annotation.RequestMethod.DELETE;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
+import static org.springframework.web.bind.annotation.RequestMethod.POST;
 
 @RestController
 @RequestMapping("/categorias")
@@ -28,8 +32,24 @@ public class CategoriaController {
         return categorias;
     }
 
-    @RequestMapping(method = DELETE)
-    public void delete(@RequestParam String descricao) {
+    @RequestMapping(method = POST, path = "/criar")
+    public void criar(@RequestBody Map<String, String> body, @RequestAttribute RoleUtilizador role) {
+        if (role != FUNCIONARIO) {
+            throw new AuthorizationServiceException("Sem autorização");
+        }
+        if (!body.containsKey("descricao")) {
+            throw new IllegalArgumentException("Dados inválidos");
+        }
+        Categoria categoria = new Categoria();
+        categoria.setDescricao(body.get("descricao"));
+        categoriaService.save(categoria);
+    }
+
+    @RequestMapping(method = DELETE, path = "/apagar")
+    public void delete(@RequestParam String descricao, @RequestAttribute RoleUtilizador role) {
+        if (role != FUNCIONARIO) {
+            throw new AuthorizationServiceException("Sem autorização");
+        }
         Optional<Categoria> optionalCategoria = categoriaService.get(descricao);
         if (optionalCategoria.isPresent()) {
             categoriaService.delete(optionalCategoria.get());
